add logging

+++ src/main/java/com/example/service/EventService.java
@@ -1,5 +1,6 @@
 package com.example.service;
 
+import com.example.entity.DefaultUser;
 import com.example.entity.Event;
 import com.example.entity.User;
 import com.example.entity.UserEvent;
@@ -10,8 +11,11 @@ import com.example.model.event.EventDto;
 import com.example.model.event.NewEventDto;
 import com.example.model.userEvent.UserEventDto;
 import com.example.repository.EventRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import java.time.LocalDate;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -22,12 +26,15 @@ public class EventService {
     private final EventRepository eventRepository;
     private final UserEventService userEventService;
     private final UserService userService;
+    private final DefaultUserService defaultUserService;
+    private static final Logger LOGGER = LoggerFactory.getLogger(EventService.class);
 
     @Autowired
-    public EventService(EventRepository eventRepository, UserEventService userEventService, UserService userService) {
+    public EventService(EventRepository eventRepository, UserEventService userEventService, UserService userService, DefaultUserService defaultUserService) {
         this.eventRepository = eventRepository;
         this.userEventService = userEventService;
         this.userService = userService;
+        this.defaultUserService = defaultUserService;
     }
 
     public EventDto add(NewEventDto eventDto, Long userId) {
@@ -40,18 +47,20 @@ public class EventService {
         userEvent.setUser(userEntity);
         userEvent.setOwner(true);
         userEventService.add(userEvent);
+        LOGGER.debug("Add event {}", eventEntity);
         return EventDto.toDto(eventEntity);
     }
 
     public Long delete(Long id, Long userId) {
         User userEntity = userService.getById(userId);
         Event eventEntity = getById(id);
-        if(eventEntity==null) throw new EventNotFoundException("Event not found");
+        if(eventEntity==null) throw new EventNotFoundException("Event not found.");
         UserEvent userEvent = userEventService.getByUserAndEvent(userEntity, eventEntity);
-        if (userEvent == null) throw new UserAccessException("User access exception.");
+        if (userEvent == null || !userEvent.getOwner()) throw new UserAccessException("User access exception.");
         List<UserEvent> userEvents = userEventService.getByEvent(eventEntity);
         userEvents.forEach(p -> userEventService.delete(p.getId()));
         eventRepository.deleteById(id);
+        LOGGER.debug("Delete event with id {}", id);
         return id;
     }
 
@@ -60,28 +69,31 @@ public class EventService {
         if (!eventDto.getStartDate().isBefore(eventDto.getEndDate())) throw new EventDateException();
         Event eventEntity = EventDto.toEntity(eventDto);
         UserEvent userEvent = userEventService.getByUserAndEvent(userEntity, eventEntity);
-        if (userEvent == null) throw new UserAccessException("User access exception.");
+        if (userEvent == null || !userEvent.getOwner()) throw new UserAccessException("User access exception.");
+        LOGGER.debug("Update event {}", eventEntity);
         return EventDto.toDto(eventRepository.save(eventEntity));
     }
 
     public EventDto subscribe(Long eventId, Long userId) {
         Event eventEntity = getById(eventId);
-        User userEntity = userService.getById(userId);
-        if(userEventService.getByUserAndEvent(userEntity, eventEntity)==null) {
-            UserEvent userEvent = new UserEvent();
-            userEvent.setEvent(eventEntity);
-            userEvent.setUser(userEntity);
-            userEvent.setOwner(false);
-            userEventService.add(userEvent);
-            return EventDto.toDto(eventEntity);
-        }
-        return null;
+        DefaultUser defaultUserEntity = defaultUserService.getById(userId);
+        if(userEventService.getByUserAndEvent(defaultUserEntity, eventEntity) != null &&
+                (LocalDate.now().getYear() - defaultUserEntity.getBDay().getYear()) < eventEntity.getAgeCensor())
+            throw new UserAccessException("User access exception.");
+        UserEvent userEvent = new UserEvent();
+        userEvent.setEvent(eventEntity);
+        userEvent.setUser(defaultUserEntity);
+        userEvent.setOwner(false);
+        userEventService.add(userEvent);
+        LOGGER.debug("Subscribe default user {} to event {}", defaultUserEntity, eventEntity);
+        return EventDto.toDto(eventEntity);
     }
 
     public Long unsubscribe(Long eventId, Long userId) {
         Event eventEntity = getById(eventId);
         User userEntity = userService.getById(userId);
         UserEvent userEvent = userEventService.getByUserAndEvent(userEntity, eventEntity);
+        LOGGER.debug("Unsubscribe user {} from event {}", userEntity, eventEntity);
         return userEventService.delete(userEvent.getId());
     }
 
@@ -89,9 +101,10 @@ public class EventService {
         Event eventEntity = getById(eventId);
         User userEntity = userService.getById(userId);
         UserEvent userEvent = userEventService.getByUserAndEvent(userEntity, eventEntity);
-        if(!userEvent.getOwner()) {
+        if(userEvent != null && !userEvent.getOwner()) {
             userEvent.setReview(review);
             userEventService.update(userEvent);
+            LOGGER.debug("Add review {} to event {}", review, eventEntity);
             return UserEventDto.toDto(userEvent);
         }
         else {
